requirements.txt
jira>=3.5.0
pandas>=2.0.0
python-dotenv>=1.0.0
# streamlit>=1.30.0       # uncomment when building dashboard
# exchangelib>=5.0.0      # uncomment when building email feature
